Update card layout
Update the card layout so the image stretches fully across the card. This
results in a nicer experience for the user.

Additionally instead of displaying book description we only show the author.
This fixes the issue of having too much information in the main page since
book description is available in the detail page.